added thumbnail on right side of screen

+++ epub_exporter.py
@@ -3,51 +3,57 @@ from ebooklib import epub
 import os
 
 
-class BookEpub:
-    def __init__(self, title):
+class Ebook:
+    def __init__(self):
         self.book_content = epub.EpubBook()
-        self.chapters = list()
-        self.isOpen = True
-    
-        # set metadata
-        self.book_content.set_identifier(title.lower().replace(" ","-"))
+        self.chapters_in_queue = list()
+
+    def init(self, title):
+        self.book_content.set_identifier(title.lower().replace(" ", "-"))
         self.book_content.set_title(title)
         self.book_content.set_language('en')
-
         self.book_content.set_cover("image.png", open('cover.png', 'rb').read())
+
+        if os.path.isfile("cover.png"):
+            os.remove("cover.png")
+
+    def init_from_file(self, path):
+        self.book_content = epub.read_epub(path)
 
     def save(self, path):
         if os.path.isfile(path):
             os.remove(path)
 
-        if self.isOpen:
-            self.isOpen = False
-            self.close_book()
-
         # write to the file
         epub.write_epub(path, self.book_content, {})
-        os.remove("cover.png")
 
     def status(self):
-        return "You have "+str(len(self.chapters))+" chapters"
-
-    def open(self, path):
-        self.isOpen = False
-        self.chapters = list()
-        self.book_content = epub.read_epub(path)
+        return "You have " + str(len(self.book_content.toc)) + " chapters loaded."
 
     def add_chapter(self, title, text):
-        chapter = epub.EpubHtml(title=title, file_name='chapter_' + str(len(self.chapters) + 1) + '.xhtml', lang='en')
+        chapter = epub.EpubHtml(title=title, file_name=title.lower().replace(" ", "-")+ '.xhtml', lang='en')
 
         chapter.content = u'<h1>' + title + '</h1>'+text
 
-        self.chapters.append(chapter)
+        self.chapters_in_queue.append(chapter)
 
-    def close_book(self):
-        for chapter in self.chapters:
+    def update_data(self):
+        chapters = reversed(self.chapters_in_queue)
+        for chapter in chapters:
+            self.book_content.add_item(chapter)
+            self.book_content.toc.append(chapter)
+        self.book_content.spine = ['nav', *chapters]
+
+        self.book_content.add_item(epub.EpubNcx())
+        self.book_content.add_item(epub.EpubNav())
+
+        self.chapters_in_queue = list()
+
+    def set_meta(self):
+        for chapter in self.chapters_in_queue:
             self.book_content.add_item(chapter)
 
-        self.book_content.toc = self.chapters
+        self.book_content.toc = self.chapters_in_queue
 
         # add default NCX and Nav file
         self.book_content.add_item(epub.EpubNcx())
@@ -61,4 +67,19 @@ class BookEpub:
         self.book_content.add_item(nav_css)
 
         # basic spine
-        self.book_content.spine = ['nav', *self.chapters]
+        self.book_content.spine = ['nav', *self.chapters_in_queue]
+
+        last = self.chapters_in_queue[-1]
+        last = last.content[0:50]
+        self.book_content.set_unique_metadata("wuxiadownloader", "lastchapter", last)
+
+        self.chapters_in_queue = list()
+
+    def get_cover(self):
+        return self.get_item_of_type(epub.EpubCover)
+
+    def get_item_of_type(self, type):
+        for item in self.book_content.items:
+            if isinstance(item, type):
+                return item
+        return None
